31-2,31-3,37-2 get Account from session

+++ backend/testhelper/src/main/java/kr/ac/ajou/da/testhelper/tests/TestsService.java
@@ -1,13 +1,12 @@
 package kr.ac.ajou.da.testhelper.tests;
 
-import java.sql.SQLException;
-import java.util.HashMap;
-import java.util.List;
-
+import kr.ac.ajou.da.testhelper.account.AccountMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import kr.ac.ajou.da.testhelper.account.AccountMapper;
+import java.sql.SQLException;
+import java.util.HashMap;
+import java.util.List;
 
 @Service
 public class TestsService {
@@ -16,7 +15,7 @@ public class TestsService {
 	@Autowired
 	private AccountMapper accountMapper;
 	
-	public List<HashMap<String, Object>> getTests(int accountId) throws SQLException {
+	public List<HashMap<String, Object>> getTests(Long accountId) throws SQLException {
 		String role = getAccountRole(accountId);
 		
 		if(role.equals("PROFESSOR")) {
@@ -28,7 +27,7 @@ public class TestsService {
 		return null;
 	}
 
-	private String getAccountRole(int accountId) throws SQLException {
+	private String getAccountRole(Long accountId) throws SQLException {
 		return accountMapper.getAccountRole(accountId);
 	}
 }
